+++ fixture-monkey-kotlin/src/test/kotlin/com/navercorp/fixturemonkey/kotlin/ExpressionGeneratorTest.kt
@@ -1,3 +1,21 @@
+/*
+ * Fixture Monkey
+ *
+ * Copyright (c) 2021-present NAVER Corp.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package com.navercorp.fixturemonkey.kotlin
 
 import org.assertj.core.api.BDDAssertions.then
@@ -175,14 +193,4 @@ class ExpressionGeneratorTest {
 
         then(actual).isEqualTo("nestedThriceDogs[1][2][2].name")
     }
-
-    data class Person(
-        val name: String?,
-        val dog: Dog,
-        val dogs: List<Dog>,
-        val nestedDogs: List<List<Dog>>,
-        val nestedThriceDogs: List<List<List<Dog>>>,
-    )
-
-    data class Dog(val name: String, val loves: List<Int>)
 }
